boolean operator feedback
