+++ utils/test/feedback/boolean_operators_test.exs
@@ -1,0 +1,5 @@
+defmodule Utils.Feedback.BooleanOperatorsTest do
+  use Utils.FeedbackCase,
+    module: Utils.Feedback.BooleanOperators,
+    file: "../exercises/boolean_operators.livemd"
+end
